requirements.txt
djangorestframework>=3.14
drf-spectacular>=0.27
Django>=5.2
djangorestframework>=3.14
psycopg2-binary>=2.9
pytest>=8.0
pytest-cov>=5.0
pytest-django>=4.0
ruff>=0.4
black>=24.3
Pillow>=10.0